+++ packages/app/src/components/catalog/CustomCatalogTable.tsx
@@ -75,7 +75,10 @@ export function CustomCatalogTable(props: CustomCatalogTableProps) {
       ),
       ...createEntitySpecificColumns(),
       hiddenColumn(GSColumnFactories.createDescriptionColumn()),
-    ];
+    ].map(column => ({
+      ...column,
+      width: 'auto',
+    }));
 
     function createEntitySpecificColumns(): TableColumn<CatalogTableRow>[] {
       const baseColumns = [
@@ -147,6 +150,7 @@ export function CustomCatalogTable(props: CustomCatalogTableProps) {
           padding: '0 18px',
         },
         columnsButton: true,
+        thirdSortClick: false,
       }}
       emptyContent={emptyContent}
     />