+++ plugins/gs/src/components/catalog/columns.tsx
@@ -41,9 +41,19 @@ export const columnFactories = Object.freeze({
         const entity1Tag = getLatestReleaseTagFromEntity(entity1);
         const entity2Tag = getLatestReleaseTagFromEntity(entity2);
 
-        return entity1Tag && entity2Tag
-          ? semver.compare(entity1Tag, entity2Tag)
-          : 0;
+        if (!entity1Tag && !entity2Tag) {
+          return 0;
+        }
+
+        if (!entity1Tag) {
+          return -1;
+        }
+
+        if (!entity2Tag) {
+          return 1;
+        }
+
+        return semver.compare(entity1Tag, entity2Tag);
       },
       customFilterAndSearch(query: string, { entity }) {
         const entityTag = getLatestReleaseTagFromEntity(entity);
@@ -66,13 +76,24 @@ export const columnFactories = Object.freeze({
       title: 'Last released',
       hidden: options.hidden,
       width: 'auto',
+      filtering: false,
       customSort({ entity: entity1 }, { entity: entity2 }) {
         const entity1Date = getLatestReleaseDateFromEntity(entity1);
         const entity2Date = getLatestReleaseDateFromEntity(entity2);
 
-        return entity1Date && entity2Date
-          ? compareDates(entity2Date, entity1Date)
-          : 0;
+        if (!entity1Date && !entity2Date) {
+          return 0;
+        }
+
+        if (!entity1Date) {
+          return 1;
+        }
+
+        if (!entity2Date) {
+          return -1;
+        }
+
+        return compareDates(entity2Date, entity1Date);
       },
       render: ({ entity }) => (
         <DateComponent
@@ -91,6 +112,7 @@ export const columnFactories = Object.freeze({
       title: 'Helm charts',
       hidden: options.hidden,
       width: 'auto',
+      filtering: false,
       customSort({ entity: entity1 }, { entity: entity2 }) {
         const entity1HelmCharts = getHelmChartsFromEntity(entity1) || [];
         const entity2HelmCharts = getHelmChartsFromEntity(entity2) || [];