+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Change the width of the catalog table columns so they evenly spread.
+- Fix catalog table 'Latest release' column sorting.
+- Fix catalog table 'Last released' column sorting.
+- Disable 'unsorted' click for catalog table order buttons.
+
 ## [0.20.0] - 2024-04-04
 
 ### Added